pyproject.toml
[project]
name = "djangoproject"
version = "0.1.0"
description = ""
authors = [
    {name = "Serhii Siuziev",email = "serg.syuzev@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "django-unfold (==0.68.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
